requirements.txt
Flask==2.1.1
requests==2.27.1
configparser~=5.2.0
tabulate~=0.9.0
python-telegram-bot==13.7